Split classes into different files
1.Move container of script object into `container/` directory
2. Move wrapper for single value into `utility.hpp`
