+++ ext/utility/include/asbind20/ext/vocabulary.hpp
@@ -4,7 +4,6 @@
 #pragma once
 
 #include <asbind20/asbind.hpp>
-#include <asbind20/container.hpp>
 
 namespace asbind20::ext
 {
