+++ include/asbind20/container/traits.hpp
@@ -1,0 +1,13 @@
+#ifndef ASBIND20_CONTAINER_TRAITS_HPP
+#define ASBIND20_CONTAINER_TRAITS_HPP
+
+#pragma once
+
+namespace asbind20::container
+{
+template <template <typename...> typename Container>
+struct container_traits
+{};
+} // namespace asbind20::container
+
+#endif
